Changes for compatibility for the Ultra96 Dualcam (u96v2_dualcam) design and PetaLinux 2022.1.

+++ recipes-app/camera-setup/files/zub1cg-sbc/load_camera_config.sh
@@ -26,38 +26,30 @@ esac
 # will prevent the module probe to fail when loading the module after start up.
 #
 # The GPIO[11] is linked the XCZU3EG's IO_L12N_AD0N_26 pin. In the FPGA design the pin IO_L12N_AD0N_26
-# is linked the output #1 of the axi_gpio_0 core at address 0xa0030000.
+# is linked the output #1 of the axi_gpio_0 core at address 0xa0000000.
 #
 # root@u96v2-sbc-dualcam-2021-1:~# cat /sys/class/gpio/gpiochip500/label
-# a0030000.gpio
+# a0000000.gpio
 # root@u96v2-sbc-dualcam-2021-1:~# cat /sys/class/gpio/gpiochip500/ngpio
 # 8
 source /usr/local/bin/gpio/gpio_common.sh
 
-BASE_A0030000=$(get_gpiochip_base a0030000)
-((AP1302_ID_GPIO=BASE_A0030000+1))
+BASE=$(get_gpiochip_base a0000000)
+((AP1302_ID_GPIO=BASE+1))
 echo "   AP1302_ID GPIO = $AP1302_ID_GPIO"
 
 echo $AP1302_ID_GPIO > /sys/class/gpio/export
 echo out > /sys/class/gpio/gpio$AP1302_ID_GPIO/direction
 echo 0 > /sys/class/gpio/gpio$AP1302_ID_GPIO/value
 
+mkdir -p /sys/kernel/config/device-tree/overlays/ap1302
 
-
-rmmod xilinx_vpss_scaler
-rmmod xilinx_csi2rxss
-# if xilinx-video driver is already binded in the kernel unbind it. This will force the xilinx-video driver
-# to recreate the V4L2 graph when we re-bind it after updating the device-tree
-if [ -e /sys/bus/platform/drivers/xilinx-video/amba_pl@0:vcap_CAPTURE_PIPELINE_v_proc_ss_scaler_0 ]; then
-	echo "amba_pl@0:vcap_CAPTURE_PIPELINE_v_proc_ss_scaler_0" > /sys/bus/platform/drivers/xilinx-video/unbind
+if [ ! -e /boot/devicetree/$1.dtbo ]; then
+	        echo "/boot/devicetree/$1.dtbo: No such file or directory"
+	        exit 2
 fi
 
-mkdir -p /sys/kernel/config/device-tree/overlays/ap1302
 cat /boot/devicetree/$1.dtbo > /sys/kernel/config/device-tree/overlays/ap1302/dtbo
-
-modprobe xilinx_vpss_scaler
-modprobe xilinx_csi2rxss
-echo "amba_pl@0:vcap_CAPTURE_PIPELINE_v_proc_ss_scaler_0" > /sys/bus/platform/drivers/xilinx-video/bind
 
 sed -i -E "s/INPUT_RESOLUTION=[0-9]+x[0-9]+/INPUT_RESOLUTION=$CAMERA_RESOLUTION/g" $(which run_1920_1080)
 sed -i -E "s/INPUT_RESOLUTION=[0-9]+x[0-9]+/INPUT_RESOLUTION=$CAMERA_RESOLUTION/g" $(which run_3840_2160)
